==> 01_unrar.ipynb <==
# Trade network analysis
**Brian Dew (brianwdew@gmail.com)**

`01_unrar.ipynb`

This script unrar's raw BACI data downloaded from CEPII into a set of csv files in a subfolder.

Data source: http://www.cepii.fr/CEPII/en/bdd_modele/presentation.asp?id=1 using version 2007 of the HS system

# patoolib has unrar, os to rename files
import patoolib
import os
os.chdir('C:/Working/trade_network/data/')
if not os.path.exists('raw/.'):
    os.makedirs('raw/.')

for y in map(str, range(2008,2015)):    # start year & end year + 1 
    filename = 'baci07_' + y           # string with name of file
    outdir = 'raw/'    
    rarfile = filename + '.rar'
    csvfile = outdir + filename + '.csv'
    csvnewname = outdir + filename + '_raw.csv'
    patoolib.extract_archive(rarfile,outdir=outdir + '.')
    os.rename(csvfile,csvnewname)
