==> 05_em-dat.ipynb <==
# Trade network analysis
**Brian Dew (brianwdew@gmail.com)**

`05_em-dat.ipynb`

Generates a csv file with the year of and country affected by large natural disasters. 

Required file:

* em-dat.csv - Raw natural disaster data from [EM-DAT: The International Disaster Database](http://www.emdat.be/) Advanced Search.

---

METODO: 


#### Libraries

import pandas as pd                                         # pandas dataframes used for convenience
import os                                                   # change current directory in next line
os.chdir('C:/Working/trade_network/data/')

read csv downloaded from em-dat. I used the advanced search on all countries years 2009-2015 and all natural disasters. I then grouped only by year and country. The resultant csv file has a column for the total deaths and a column labeled occurence which contains the number of disasters. I'm interested in really large disasters, which must be large enough, or few enough, to have a major economic impact through decreased exports. I therefore have a cutoff below for the average number of deaths per occurance. I find that large countries have many disasters and therefore a cutoff, of say 100, in absolute terms will add the US, India, and China to the list for each year. There are a few years where these countries had major natural disasters that I seek to include in this measure. The cutoff can be adjusted to check for robustness.

df = pd.read_csv('em-dat.csv', header=0, index_col='year')
em_countries = {}
for y in range(2008,2015):
    em_countries[y] = df[ (df['Total deaths'] / df['occurrence'] ) > 39.9].loc[y]   # CUTOFF
    em_countries[y]['y'] = y                                            
emergency_countries = pd.concat(em_countries, axis=0).reset_index()[['y','iso']]
emergency_countries['s'] = 1     
# Create a dictionary of iso2 country codes, Namibia is interpreted as NaN without the keep_default_na
iso2 = pd.read_csv('country_code_baci07.csv', index_col='iso3', keep_default_na=False, na_values=[''])['iso2'].to_dict()
emergency_countries['iso'] = emergency_countries['iso'].apply(lambda x: iso2.get(x,x)) # replace country names
# to merge and create dummy
emergency_countries.to_csv('emergency_countries.csv', index=None)    # save as csv file
