==> 10_tables.ipynb <==
# Trade network analysis
**Brian Dew (brianwdew@gmail.com)**

`10_tables.ipynb`

Generates LaTeX tables with summary statistics on:

Required file:

* `panel.csv` - created by `09_panel.ipynb`

---

METODO: 

Annotate and expand

#### Libraries

import pandas as pd                                         # pandas dataframes used for convenience
import os                                                   # change current directory in next line
import numpy as np

os.chdir('C:/Working/trade_network/data/')

#### Summary of largest exporters

df = pd.read_csv('panel.csv', index_col=['iso2', 'date'], parse_dates=['date'], header=0)
df['x'] = df['x'] / 1000000000
df['tot_ch_x'] = df['x_def'].groupby(df.index.get_level_values(0)).diff(periods=8) 
df['ann_ch_x'] = df['x_def'].groupby(df.index.get_level_values(0)).diff(periods=1) 
std_dev = df['ann_ch_x'].groupby(df.index.get_level_values(0)).std()

df1 = df.query('date == "2015-01-01"').sort_values('x_ndef', ascending=False)[['full_name', 'x_ndef', 'tot_ch_x']]
df1['x_ndef'] = df1['x_ndef'].apply(float) / 1000000000
df1 = df1.join(std_dev)
df1 = np.round(df1, decimals=2)
df1.to_csv('export_country_table.csv')
print df1.to_latex(index=False)
