==> 12_product_network.ipynb <==
# Trade network analysis
**Brian Dew (brianwdew@gmail.com)**

`12_product_network.ipynb`

For `prod` product, calculate the eigenvector centralities, and shares of total trade by country. Save as csv in `data/prod`

METODO:


# LIBRARIES
import pandas as pd 
import numpy as np #for eigenvector calc
import networkx as nx 
import pickle #save graphs
%matplotlib inline

# FILEPATHS
path = r'C:/Working/trade_network/data/clean'
dloc = r'C:/Working/trade_network/data' #destination of final file
pickle_loc = r'C:/Working/trade_network/data/tmp'

Build network and complete network calculations

# Select product
prod = 271121  # product id to use

frame = []
# 
# 1. Build the network
#
    # Generate network G for the product above for all years 
for y in range(2008,2015):
    csv_file = path + '/baci07_' + str(y) + '_clean.csv'
    df  = pd.read_csv(csv_file, index_col='hs6', header=0)
    G   = nx.from_pandas_dataframe(df[df.index == prod], 'i', 'j', ['v', 'q'], nx.DiGraph())
#
# 2. Calculate indicators for the network
#
    # Calculate centrality measures
    outdeg_cent = nx.out_degree_centrality(G)
    indeg_cent =  nx.in_degree_centrality(G)
    outdeg = G.out_degree(weight='q')
    indeg = G.in_degree(weight='q')
    cim = nx.eigenvector_centrality_numpy(G, weight='q')
    cex = nx.eigenvector_centrality_numpy(G.reverse(), weight='q')
        
    # Set the indicators as node attributes
    indicators = {
        'c_od': outdeg_cent,   # eigenvector centrality of importers
        'c_id': indeg_cent,   # eigenvector centrality of exporters       
        'ecim': cim,   # eigenvector centrality of importers
        'ecex': cex,   # eigenvector centrality of exporters
        'outd': outdeg,# total exports for country
        'indg': indeg,  # total imports for country             
        
    }
    for k, v in indicators.iteritems():
        nx.set_node_attributes(G, k, v)
    
    # Temporarily store the network for that year
    pickle.dump(G, open(pickle_loc + '/graph_'+ str(y) + '_' + str(prod) + '.txt', 'w'))
    
#
# 3. Build dataframe and merge network, indicators, and imfdata
#
    df = pd.concat([pd.Series(j[1]) for j in indicators.iteritems()], axis=1, keys=[j[0] for j in indicators.iteritems()])
    df['year'] = pd.to_datetime(str(y)).year
    df['tot'] = G.size(weight='q') 
    df['shX'] = df['outd'] / df['tot']
    df['shM'] = df['indg'] / df['tot']
    df.index.names=['iso2']
    
    frame.append(df) ## store dataframes in list

combined = pd.concat(frame, axis=0)
combined.reset_index().set_index(['iso2','year']).to_csv('data/prod/final_'+ str(prod) + '.csv')

combined = pd.read_csv('data/prod/final_271111.csv', index_col=['iso2'])
combined[combined.index == 'JP']
